+++ src/main/java/com/sidenow/global/config/redis/service/RefreshTokenServiceImpl.java
@@ -7,7 +7,6 @@
 //import com.sidenow.global.config.redis.dto.RefreshTokenDto;
 //import com.sidenow.global.config.redis.exception.NotFoundRefreshToken;
 //import com.sidenow.global.config.redis.repository.RedisRepository;
-//import com.sidenow.global.dto.TokenInfoResponse;
 //import jakarta.transaction.Transactional;
 //import lombok.RequiredArgsConstructor;
 //import lombok.extern.slf4j.Slf4j;
